// models/CustomData.ts
export type CustomData = {
  name?: string;
  address?: string;
  email?: string;
  phone?: string;
  consultation?: {
    userNote: string;
    designerNote: string;
  };
  revisionOne?: {
    userNote: string;
    designerNote: string;
  };
  revisionTwo?: {
    userNote: string;
    designerNote: string;
  };
  revisionThree?: {
    userNote: string;
    designerNote: string;
  };
};